Separate editing and deploy workflow
This is to avoid running them in parallel as it results into downloading models multiple times. If we keep them separate, we can run them sequentially when building the app
